Add support for imagebuilder as an alternative builder for docker images
[imagebuilder](github.com/openshift/imagebuilder) is an alternative to `docker build` for building secure(A volume mount inside a build context, no need to leave secrets inside the docker image/metadata in case you have `bundle install` or `npm install` or etc. to install packages hosted in private Git repos) and small docker images(w/ squashing) in a faster way(No context upload).

A new setting key `image_builder` is added under `containers` in service.yaml, which can be omitted or set to either `docker` or `imagebuilder`.

If the value was omitted or set to `docker`, `docker build` is used for building docker images. This is the existing and default behavior before/after this change.

If the value was set to `imagebuilder`, the `imagebuilder` is used instead.

The only gotcha is that to support `--build-arg`s for `imagebuilder`, you need https://github.com/openshift/imagebuilder/pull/58 until it gets merged.